# Improve MetaMember reflection performance 