… OpenRiaServices.DomainServices.Client/Framework/Silverlight/Data/EntityConflict.cs
… OpenRiaServices.DomainServices.Client/Framework/Silverlight/Data/EntityVisitor.cs
… OpenRiaServices.DomainServices.Client/Framework/Silverlight/Data/MetaType.cs
… OpenRiaServices.DomainServices.Client/Framework/Silverlight/Data/ObjectStateUtility.cs
… OpenRiaServices.DomainServices.Client/Framework/Silverlight/Data/ValidationUtilities.cs
… OpenRiaServices.DomainServices.Client/Test/Desktop/OpenRiaServices.DomainServices.Client.Test.csproj
… OpenRiaServices.DomainServices.Server/Framework/Data/MetaType.cs