Improve reflection logic for performance improvements in a number of scenarios.

Examples of affected changes includes 
* Load
* Submit
* ToString 
* Builtin fallback  implementation of GetIdentity
* Validation specifically on entity level

Approximate numbers for different methods when processing 1000 of the "Citites" entities found in the test project. The timings include client only processing (excluding serialization/deserialization) and timings are taken from a single run of the Benchmarks (https://github.com/OpenRIAServices/Benchmarks) project for master compared to this branch.

|Metod| Average approximate speedup | % or original | 
|---|--|--|
|Add|+27%|79%|
|Attach|+20%|84%|
|Submit|+169%|37%|
|Load|158%|39%|
|Modify attached entity|21%|82%|
|Total time (excluding test setup)|47%|68%|