@Daniel-Svensson 2018-02-20
TODO:
[ ] Fix regression in test _NamedUpdate_CustomValidation_CommonProperties_